+++ NEWS.md
@@ -1,5 +1,13 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# cynkrathis 0.3.2 (2021-06-17)
+
+- `deploy_minicran_package()` now also works if no global git config was set by copying the projects git config to the temporary drat directory (#21)
+- `deploy_minicran_package()` uses a fresh directory for each clone task now instead of reusing the previous one which might lead to conflicts (#22)
+- Add `renv_install_local()`, an {renv} helper which helps building local packages by moving them to the shared {renv} cache location for local packages and installing them - all in one step
+- Add `renv_donwgrade()`, a helper for downgrading to old RSPM snapshots
+
+
 # cynkrathis 0.3.1 (2021-06-01)
 
 - `deploy_minicran_package()`: compatibility with drat 0.2.0